C51-386: Segregate backstopJS setup from gulpfile
